Change fallible return type to Result union
