+++ src/lib.rs
@@ -354,10 +354,10 @@ mod assert;
 
 mod cxx_string;
 mod error;
-mod exception;
 mod gen;
 mod opaque;
 mod paths;
+mod result;
 mod rust_str;
 mod rust_string;
 mod syntax;
@@ -371,8 +371,8 @@ pub use cxxbridge_macro::bridge;
 // Not public API.
 #[doc(hidden)]
 pub mod private {
-    pub use crate::exception::{r#try, Error};
     pub use crate::opaque::Opaque;
+    pub use crate::result::{r#try, Result};
     pub use crate::rust_str::RustStr;
     pub use crate::rust_string::RustString;
     pub use crate::unique_ptr::UniquePtrTarget;
